+++ package/assets/Card-BvLj4L6F-hBAEBJnk.js
@@ -1,0 +1,1 @@
+import{e as o,d,h as p,p as r,ak as n,y as s}from"./index-CFUzHmV3.js";var i=o(d({__name:"Card",props:{padding:{type:Boolean,default:!0}},setup(a){return(e,t)=>(p(),r("div",{class:s(["phlix-card",{"phlix-card--no-padding":!a.padding}])},[n(e.$slots,"default",{},void 0)],2))}}),[["__scopeId","data-v-0e99c3b2"]]);export{i as o};